src/requirements.txt
python-dotenv == 1.0.1
langchain == 0.3.0
azure-search-documents == 11.6.0b2
